Move App.js to webapp folder and modify index.js
