+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App.js';
+import App from './main/webapp/App.js';
 
 ReactDOM.render(
   <React.StrictMode>
